pyproject.toml
[tool.poetry]
name = "amazing_hunting"
version = "0.1.0"
description = ""
authors = ["Dmitriy Lahmenev <lahmenev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
psycopg2 = "^2.9.3"
Pillow = "^9.2.0"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
